+++ packages/models/src/lib/content/index.ts
@@ -1,5 +1,3 @@
 export { ContentModel } from './content.model';
-export { NewsContentModel } from './news-content.model';
 
-export { NewsCategory } from './news-category.enum';
-export { NewsForm } from './news-form.model';
+export * from './news';
