+++ packages/models/src/lib/content/news/index.ts
@@ -1,0 +1,3 @@
+export { NewsContentModel } from './news-content.model';
+export { NewsCategory } from './news-category.enum';
+export { NewsForm } from './news-form.model';
